refactor(workflows): replace pluginCtx global with accessor + dedupe search-sync
Phase C1 of the workflows/index.ts split. Removes the mutable `pluginCtx`
module global and collapses the duplicated search helpers.

- lib/plugin-context.ts — setWorkflowPluginContext / getWorkflowPluginContext.
  Route closures are built at module-load (populateWorkflowRoutes runs at
  import, before activate) so they can't capture ctx directly; they read it
  lazily through this accessor at request time. activate sets it once.
- lib/search-sync.ts — one ctx-injected set of instanceToSearchDoc /
  definitionToSearchDoc / indexInstance (previously duplicated module-scope
  reading the pluginCtx global AND activate-scope closing over ctx), plus a
  new indexDefinition for the availability route. indexInstance reads the
  accessor and no-ops before activate, matching the old `if (!pluginCtx)`
  guard.

index.ts no longer declares pluginCtx at all: the start route reads
getWorkflowPluginContext(), activate calls setWorkflowPluginContext(ctx), and
the registerFileBackedContentType block imports the search-doc builders.

Behavior-preserving: the accessor holds the last-activated ctx exactly as the
old module global did; the unified indexInstance keeps the guard.

Verified: typecheck + eslint + workflows suite (420/0) + full suite (5124/0)
+ binary build + isolated-home boot smoke (bakin_workflows content type
registers; GET /definitions and /search return 200).
